First break commit.
Made the top floor of the level with no interaction completed yet.

First person character controller created with jump, walk, run movement integrated into it.

Need to create ability to interact with the "interactable" objects to be able to complete actions with them.

Need to create inventory system so you can only interact with the destructables when you have the correct tools.