Added in cargo boxes and started game manager.
Cargo boxes can contain items or be empty.

Player has an inventory that has the ability to contain a list of items and check for that item.

NEED TO ADD:
Need to add system to pick up an item.

Need to add system for the objectives to check for correct items to complete game.